Move d.ts file location. Confirm webpack5 bug.
